Refactoriza la ruta del archivo constants.dart
