+++ lib/components/task_card_helper.dart
@@ -1,4 +1,4 @@
-import 'package:dcristaldo/constants.dart';
+import 'package:dcristaldo/constants/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:dcristaldo/domain/task.dart';
 import 'package:dcristaldo/helpers/common_widgets_helper.dart'; // Importa CommonWidgetsHelper
