+++ lib/views/detail_screen.dart
@@ -1,5 +1,5 @@
 import 'package:dcristaldo/api/services/task_service.dart';
-import 'package:dcristaldo/constants.dart';
+import 'package:dcristaldo/constants/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:dcristaldo/domain/task.dart';
 import 'package:dcristaldo/helpers/common_widgets_helper.dart'; // Importa CommonWidgetsHelper
