+++ lib/views/noticia_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:dcristaldo/data/noticia_repository.dart';
 import 'package:dcristaldo/components/noticia_card.dart';
 import 'package:dcristaldo/views/base_screen.dart';
-import 'package:dcristaldo/constants.dart';
+import 'package:dcristaldo/constants/constants.dart';
 import 'package:dcristaldo/domain/noticia.dart';
 import 'package:intl/intl.dart';
 import 'package:dcristaldo/exceptions/api_exception.dart';
