+++ lib/views/result_screen.dart
@@ -1,4 +1,4 @@
-import 'package:dcristaldo/constants.dart';
+import 'package:dcristaldo/constants/constants.dart';
 import 'package:flutter/material.dart';
 import 'package:dcristaldo/views/start_screen.dart';
 import 'package:dcristaldo/helpers/common_widgets_helper.dart';
